Modified doodling canvas.
Instead of having a blank canvas where the user can draw lines (Microsoft Paint-like program), the canvas now consists of cells that start white and turn black if the user clicks and drags the mouse onto them. This allows the to-be-implemented genetic algorithm to be much faster, only having to check differences between cells and not pixels when calculating the fitness of a given individual.